26/2/17: Added remote code which atm reads 4 potentiometers(thrust,pitch,ywa,roll)
Included transceivers headers and source codes along with this
controller code in same directory(sort off). Src folder which has
rtc12.h and rtc12.c which contains the main contents of initializing and
setting up the transceivers.  Still yeet to integrate with the
transceiver functions but the resources are available to use them