wrote a bit of downlink code as well in base_station
but there is a the uplink problem where there is a handshake issue where
the transmission of successive data will happen fast it wont give
receiver time to read from buffer process it and emty the recive buffer
in the drone to allow it to get ready to receive the next transmitted
value. Either do delay in base station when u transmitted a data to give
time for the receiver to get ready or do handshake.