2/27/17 report
Only focused on interface wit pc trying to spend a lot of time testing
my uart to communicate with my pc and then got the comms between pc and
ilmatto to work.
The attempted to enter floating point values which had to do manual
conversion to transmittable form (int or uint16_t).
We will most likely stick between between 0-10.24 so its programmed it
only take this range. These values get converted to int ish and that
gets  passed to joels functions wich takes taakes it and converts and
splits its 10-bit equivlent to the 8 bit at data packet and 2 bit to the
datatype.